+++ packages/camera/camera/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.9.4+13
+
+* Updates iOS camera's photo capture delegate reference on a background queue to prevent potential race conditions, and some related internal code cleanup.
+
 ## 0.9.4+12
 
 * Skips unnecessary AppDelegate setup for unit tests on iOS.